# registry/api_docs/swagger.yaml
definitions:
  http.CreateBlockRequest:
    properties:
      catalog:
        type: object
      docker_image:
        type: string
      documentation_url:
        type: string
      kind:
        type: string
      name:
        type: string
      specification:
        $ref: '#/definitions/http.Specification'
      thumbnail_url:
        type: string
      type:
        type: string
      version:
        type: string
    type: object
  http.CreateBlockResponse:
    properties:
      catalog:
        type: object
      created_at:
        type: string
      docker_image:
        type: string
      documentation_url:
        type: string
      id:
        type: string
      name:
        type: string
      specification:
        $ref: '#/definitions/http.Specification'
      thumbnail_url:
        type: string
      type:
        type: string
      updated_at:
        type: string
      version:
        type: string
    type: object
  http.GetBlockByNameAndVersion:
    properties:
      created_at:
        type: string
      docker_image:
        type: string
      documentation_url:
        type: string
      id:
        type: string
      kind:
        type: string
      name:
        type: string
      specification:
        $ref: '#/definitions/http.Specification'
      thumbnail_url:
        type: string
      type:
        type: string
      updated_at:
        type: string
      version:
        type: string
    type: object
  http.GetLatestBlock:
    properties:
      created_at:
        type: string
      docker_image:
        type: string
      documentation_url:
        type: string
      id:
        type: string
      kind:
        type: string
      name:
        type: string
      specification:
        $ref: '#/definitions/http.Specification'
      thumbnail_url:
        type: string
      type:
        type: string
      updated_at:
        type: string
      version:
        type: string
    type: object
  http.RegistryResponse-any:
    properties:
      data: {}
      error:
        type: string
    type: object
  http.RegistryResponse-http_CreateBlockResponse:
    properties:
      data:
        $ref: '#/definitions/http.CreateBlockResponse'
      error:
        type: string
    type: object
  http.RegistryResponse-http_GetBlockByNameAndVersion:
    properties:
      data:
        $ref: '#/definitions/http.GetBlockByNameAndVersion'
      error:
        type: string
    type: object
  http.RegistryResponse-http_GetBlocksByNameResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/http.GetBlockByNameAndVersion'
        type: array
      error:
        type: string
    type: object
  http.RegistryResponse-http_GetLatestBlocksResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/http.GetLatestBlock'
        type: array
      error:
        type: string
    type: object
  http.Specification:
    properties:
      apiVersion:
        type: string
      author:
        type: string
      build:
        type: object
      env:
        type: object
      gpu:
        type: boolean
      inputs:
        type: object
      outputs:
        type: object
      parameters:
        type: object
      tags:
        items:
          type: string
        type: array
    type: object
host: localhost:8080
info:
  contact: {}
  description: HTTP API for the Clay block registry.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Clay Registry
  version: "1.0"
paths:
  /v1/blocks:
    get:
      description: Return all the blocks with latest version.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RegistryResponse-http_GetLatestBlocksResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.RegistryResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.RegistryResponse-any'
      summary: Get the latest blocks from the registry
      tags:
      - Block
    post:
      consumes:
      - application/json
      description: publish the specific version of the block.
      parameters:
      - description: Block to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.CreateBlockRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.RegistryResponse-http_CreateBlockResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.RegistryResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.RegistryResponse-any'
      summary: Publish the block to clay registry
      tags:
      - Block
  /v1/blocks/{name}:
    get:
      description: Returns all the versions of the block name.
      parameters:
      - description: Name of the block
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RegistryResponse-http_GetBlocksByNameResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.RegistryResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.RegistryResponse-any'
      summary: Get block by name.
      tags:
      - Block
  /v1/blocks/{name}/versions/{version}:
    get:
      description: Returns the block for the given version and name.
      parameters:
      - description: Name of the block
        in: path
        name: name
        required: true
        type: string
      - description: Version of the block
        in: path
        name: version
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RegistryResponse-http_GetBlockByNameAndVersion'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.RegistryResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.RegistryResponse-any'
      summary: Get block by name and version.
      tags:
      - Block
swagger: "2.0"